github/workflows: remove pip usage in mingw CI
This was only needed because the mingw CI used to run on Ubuntu 20.04
which had a version of meson too old for mpv. This hasn't been the case
since we switched to 22.04 in abcd29b3284199d1dc6707fe9efff2793e90977f
and can now just use the package manager version.