wayland: unbreak runtime geometry/autofit changes
Regressed from 4a6f4f8dcef5a0868320a89f8f07063f52a337f3. The geometry
needs to be explictly recalculated now. Change up this function a little
bit also give it the ability to directly perform a resize after the
fact. This is a common workflow and we'll be using it in the next
commit.